Make the LikeButton its own View so we can re-use it in the CommentCard
