+++ Circles/Views/Timeline/MessageCard.swift
@@ -244,38 +244,6 @@ struct MessageCard: MessageView {
         }
     }
     
-
-    @ViewBuilder
-    var likeButton: some View {
-        
-        let likers = message.reactions["❤️"] ?? []
-        let iLikedThisMessage = likers.contains(message.room.session.creds.userId)
-        
-        AsyncButton(action: {
-            // send ❤️ emoji reaction if we have not sent it yet
-            // Otherwise retract it
-            if iLikedThisMessage {
-                // Redact the previous reaction message
-                try await message.sendRemoveReaction("❤️")
-            } else {
-                // Send the reaction
-                try await message.sendReaction("❤️")
-            }
-        }) {
-            HStack(alignment: .center, spacing: 2) {
-                let icon = iLikedThisMessage ? SystemImages.heartFill : SystemImages.heart
-                let color = iLikedThisMessage ? Color.accentColor : Color.primary
-                Image(systemName: icon.rawValue)
-                    .frame(width: 20, height: 20)
-                    .foregroundColor(color)
-                Text("\(message.reactions["❤️"]?.count ?? 0)")
-            }
-            .font(footerFont)
-            .foregroundColor(footerForegroundColor)
-        }
-        .disabled(!iCanReact)
-    }
-    
     @ViewBuilder
     var commentsButton: some View {
         Button(action: {
@@ -355,7 +323,7 @@ struct MessageCard: MessageView {
 
         HStack(alignment: .top) {
             HStack(alignment: .top, spacing: 24) {
-                likeButton
+                LikeButton(message: message)
                 
                 if !isThreaded {
                     commentsButton
